@@ -4,7 +4,7 @@ cmake_minimum_required(VERSION 2.6)
 
 find_package(CGAL REQUIRED)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 
 if (CGAL_FOUND) 
 
