@@ -1,16 +1,18 @@
 #ifndef CGAL_LEVEL_OF_DETAIL_TRAITS_H
 #define CGAL_LEVEL_OF_DETAIL_TRAITS_H
 
+// New CGAL includes.
 #include <CGAL/Loader/Level_of_detail_loader.h>
 
 namespace CGAL {
 
 	namespace LOD {
 
-		template<class KernelTraits, class Container>
+		template<class KernelTraits, class OutputContainer>
 		struct Level_of_detail_traits {
 
-			typedef KernelTraits Kernel;
+			typedef KernelTraits 							  Kernel;
+			typedef OutputContainer                           Container;
 			typedef Level_of_detail_loader<Kernel, Container> Loader;
 
 		};
