@@ -1,6 +1,7 @@
 #ifndef CGAL_MYLOG_H
 #define CGAL_MYLOG_H
 
+// STL includes.
 #include <string>
 #include <sstream>
 #include <fstream>
@@ -8,46 +9,49 @@
 
 namespace CGAL {
 
-	class Mylog {
+	namespace LOD {
 
-	public:
-		std::string state() const {
-			return "ok";
-		}
+		class Mylog {
 
-		std::string data() const {
-			return out.str();
-		}
-
-		void clear() {
-			out.str(std::string());
-		}
-
-		template<typename T>
-		void append(const T &token) {
-			out << token;
-		}
-
-		bool save(const std::string &fileName) const {
-
-			const std::string path = "/Users/danisimo/Documents/pipeline/logs/";
-			const std::string finalPath = path + fileName + ".log";
-
-			std::ofstream file(finalPath.c_str(), std::ios_base::out);
-
-			if (!file) {
-				std::cerr << "\nERROR: Error saving log file with the name " << fileName << "\n" << std::endl;
-				return false;
+		public:
+			std::string state() const {
+				return "ok";
 			}
 
-			file << data() << std::endl;
-			file.close();
+			std::string data() const {
+				return out.str();
+			}
 
-			return true;
-		}
+			void clear() {
+				out.str(std::string());
+			}
 
-		std::stringstream out;
-	};
+			template<typename T>
+			void append(const T &token) {
+				out << token;
+			}
+
+			bool save(const std::string &fileName) const {
+
+				const std::string path = "/Users/danisimo/Documents/pipeline/logs/";
+				const std::string finalPath = path + fileName + ".log";
+
+				std::ofstream file(finalPath.c_str(), std::ios_base::out);
+
+				if (!file) {
+					std::cerr << "\nERROR: Error saving log file with the name " << fileName << "\n" << std::endl;
+					return false;
+				}
+
+				file << data() << std::endl;
+				file.close();
+
+				return true;
+			}
+
+			std::stringstream out;
+		};
+	}
 }
 
 #endif // CGAL_MYLOG_H
